Progetto/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==3.0.3
alembic==1.13.2
Flask-Migrate==4.0.7

